Don't use aliases when syncing assets.

+++ drupal/Sync.py
@@ -22,22 +22,28 @@ def backup_db(shortname, staging_branch, stage_drupal_root):
 
 # Sync uploaded assets from production to staging
 @task
-def sync_assets(orig_host, shortname, staging_shortname, staging_branch, prod_branch, config, remote_files_dir=None, staging_files_dir=None, sync_dir=None):
+def sync_assets(orig_host, shortname, staging_shortname, staging_branch, prod_branch, config, app_dir, remote_files_dir=None, staging_files_dir=None, sync_dir=None):
   # Switch the credentials with which to connect to production
   env.host = config.get(shortname, 'host')
   env.user = config.get(shortname, 'user')
   env.host_string = '%s@%s' % (env.user, env.host)
+
+  prod_site_exists = common.Utils.get_previous_build(shortname, prod_branch, 0)
+  if prod_site_exists is None:
+    raise SystemError("Couldn't find this site with Drush alias %s_%s in production in order to sync its files to staging! Aborting." % (shortname, prod_branch))
+
+  prod_drupal_root = prod_site_exists + '/' + app_dir
 
   if sync_dir is None:
     sync_dir = '/tmp'
  
   # Sync down the assets to the Jenkins machine, before sending them upstream to the staging server.
   print "===> Finding the remote files directories to rsync from..."
-  if run('drush @%s_%s dd files' % (shortname, prod_branch)).failed:
+  if run('cd %s && drush dd files' % prod_drupal_root).failed:
     raise SystemError("Couldn't find this site with Drush alias %s_%s in production in order to sync its assets to staging! Aborting." % (shortname, prod_branch))
   else:
     if remote_files_dir is None:
-      remote_files_dir = run("drush @%s_%s dd files" % (shortname, prod_branch))
+      remote_files_dir = run("cd %s && drush dd files" % prod_drupal_root)
     local("rsync -e 'ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no' -aHPv %s@%s:%s/ %s/%s_drupal_files/" % (env.user, env.host, remote_files_dir, sync_dir, shortname))
   
   # Switch the host to the staging server, it's time to send the assets upstream
